+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   `return self.foo()` => `return await self.foo()`
 - Deprecations: Emit warnings when using the Legacy Alerting APIs. Stop receiving
   exceptions when using the Library Elements API.
+- Smart query: Grafana 13 only supports addressing data sources by UID
 
 ## 5.0.2 (2026-02-15)
 - Fix: In the alert provisioning subsystem, added `disable_provenance` parameter